Github action to deploy a Kaizo microservice to a GKE cluster (#1)
Please notice that this is a public repository! that is because Github Actions currently doesn't support loading actions from private repositories, or at least not without a some tricky workarounds. I do not think there is anything proprietary in this repo and the Github action is quite specific for Kaizo. That said, it could serve as a public OSS example of how to do this kind of stuff

This action uses the following tools:
- The Google Cloud SDK
- Kustomize
- kubectl
- Bash, with envsubst for some very basic templating to set the project name in the base resource files

Interesting files:
- The entrypoint.sh file does most of the work
- The action.yaml file contains all the metadata for how this action should be used
- The Dockerfile is a custom version of the Google cloud-sdk for Alpine with kustomize and kubectl added.

Example usage:
- The PR in dojo-master that uses this action
- A successful build of that PR (check its full output log for all the details)